cleanup + fix tests
