--- src/DynamicVehicleScheduling/policy/anticipative_policy.jl
@@ -1,17 +1,0 @@
-"""
-$TYPEDEF
-
-Anticipative policy for the Dynamic Vehicle Scheduling Problem.
-"""
-struct AnticipativeVSPPolicy <: AbstractDynamicPolicy end
-
-"""
-$TYPEDSIGNATURES
-
-Apply the anticipative policy to the environment.
-"""
-function run_policy!(
-    ::AnticipativeVSPPolicy, env::DVSPEnv, scenario=env.scenario; model_builder=highs_model
-)
-    return generate_anticipative_solution(env, scenario; model_builder, reset_env=true)
-end
